Add demoScripts. Fix crashing Cypher_Main.json
The automatic demo.html generator pastes all of the .json data into the html file and does not work in some instances. I'm now using the .js files in /demoScripts in demo.html.

Cypher_Main.json now plays all the way through with some remaining issues. I use a write-on effect to make the letters red and that is causing performance issues. I'll be switching to a cross dissolve.